apps/prefect-etl/pyproject.toml
[project]
name = "prefect-etl"
version = "0.1.0"
description = "ETL workflows using Prefect"
requires-python = ">=3.9"
dependencies = [
    "prefect>=2.19.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "neo4j>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]